permissao usuario & logout

+++ src/main/java/com/totem/service/AtividadeService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import com.totem.dto.AtividadeDTO;
 import com.totem.entity.Atividade;
 import com.totem.entity.SubAtividade;
+import com.totem.exception.CustomErrorException;
 import com.totem.repository.AtividadeRepository;
 
 @Service
@@ -17,27 +18,44 @@ public class AtividadeService {
 	
 	
 	@Autowired
+	UsuarioService usuarioService;
+	
+	@Autowired
     private SubAtividadeService subAtividadeService;
 
 	@Autowired
     private AtividadeRepository atividadeRepository;
 	
+	private static final String ERRO_PERMISSAO = "Usuário em premissão";  
+	
 
-	public List<Atividade> listar() {
+	public List<Atividade> listar(String emailUsuario) {
+		
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
+		
 		return atividadeRepository.findAll();
 	}
 	
-	public Atividade findById (Long id) {
+	public Atividade findById (Long id, String emailUsuario) {
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		return atividadeRepository.findById(id).get();
 	}
 	
-	public Atividade salvar(AtividadeDTO atividadeDTO) {
+	public Atividade salvar(AtividadeDTO atividadeDTO, String emailUsuario) {
 		
-		List<SubAtividade> subAtividadeList = new ArrayList<SubAtividade>();
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
+		
+		List<SubAtividade> subAtividadeList = new ArrayList<>();
 		Atividade atividade = new Atividade();
 		
 		for (Long codSubAtividade : atividadeDTO.getSubAtividadeList()) {
-			subAtividadeList.add(subAtividadeService.findById(codSubAtividade));
+			subAtividadeList.add(subAtividadeService.findById(codSubAtividade, emailUsuario));
 		}
 		
 		BeanUtils.copyProperties(atividadeDTO, atividade);
@@ -48,7 +66,12 @@ public class AtividadeService {
 		return atividade;
 	}
 	
-	public Atividade delete(Long id) {
+	public Atividade delete(Long id, String emailUsuario) {
+		
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
+		
 		Atividade atividade = atividadeRepository.findById(id).get();
 		atividadeRepository.delete(atividade);
 		return atividade;
