+++ src/main/java/com/totem/service/EtapaService.java
@@ -6,16 +6,25 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.totem.entity.Etapa;
+import com.totem.exception.CustomErrorException;
 import com.totem.repository.EtapaRepository;
 
 @Service
 public class EtapaService {
 
 	@Autowired
+	UsuarioService usuarioService;
+	
+	@Autowired
     private EtapaRepository etapaRepository;
 	
+	private static final String ERRO_PERMISSAO = "Usuário em premissão";  
+	
 
-	public List<Etapa> listar() {
+	public List<Etapa> listar(String emailUsuario) {
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		return etapaRepository.findAll();
 	}
 	
@@ -23,13 +32,19 @@ public class EtapaService {
 		return etapaRepository.findById(id).get();
 	}
 	
-	public Etapa salvar(Etapa etapa) {
+	public Etapa salvar(Etapa etapa, String emailUsuario) {
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		
 		etapaRepository.save(etapa);
 		return etapa;
 	}
 	
-	public Etapa delete(Long id) {
+	public Etapa delete(Long id, String emailUsuario) {
+		if(!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		Etapa etapa = etapaRepository.findById(id).get();
 		etapaRepository.delete(etapa);
 		return etapa;
