+++ src/main/java/com/totem/service/SubAtividadeService.java
@@ -6,30 +6,47 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.totem.entity.SubAtividade;
+import com.totem.exception.CustomErrorException;
 import com.totem.repository.SubAtividadeRepository;
 
 @Service
 public class SubAtividadeService {
+	
+	private static final String ERRO_PERMISSAO = "Usuário em premissão";  
 
 	@Autowired
-    private SubAtividadeRepository subAtividadeRepository;
-	
+	UsuarioService usuarioService;
 
-	public List<SubAtividade> listar() {
+	@Autowired
+	private SubAtividadeRepository subAtividadeRepository;
+
+	public List<SubAtividade> listar(String emailUsuario) {
+		if (!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		return subAtividadeRepository.findAll();
 	}
-	
-	public SubAtividade findById (Long id) {
+
+	public SubAtividade findById(Long id, String emailUsuario) {
+		if (!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		return subAtividadeRepository.findById(id).get();
 	}
-	
-	public SubAtividade salvar(SubAtividade subAtividade) {
-		
+
+	public SubAtividade salvar(SubAtividade subAtividade, String emailUsuario) {
+		if (!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
+
 		subAtividadeRepository.save(subAtividade);
 		return subAtividade;
 	}
-	
-	public SubAtividade delete(Long id) {
+
+	public SubAtividade delete(Long id, String emailUsuario) {
+		if (!usuarioService.isAdm(emailUsuario)) {
+			throw new CustomErrorException(ERRO_PERMISSAO);
+		}
 		SubAtividade subAtividade = subAtividadeRepository.findById(id).get();
 		subAtividadeRepository.delete(subAtividade);
 		return subAtividade;
